허브간 이동 정보 생성, 조회, 수정, 삭제 API 구현 (#48)
* refactor: Hub.create() ↔ Command 의존성 제거

* fix: longitude, latitude Long -> Double로 변경

* feat: 허브 서비스 common module 추가

* test: registerHubRouteInfo 단위테스트

* feat: 허브간 이동 정보 생성 API 구현

* feat: 허브간 이동 정보 조회 API 구현

* feat: 허브간 이동 정보 수정 API 구현

* refactor: 허브 수정 API 예외처리 BusinessException으로 변경

* feat: 허브간 이동 정보 삭제 API 구현

* refactor: 허브 등록 API 예외처리 BusinessException으로 변경

* chore: api별 주석 추가

* feat: 허브간 이동 정보 생성 스케줄러 구현

* feat: 허브간 이동 정보 생성 스케줄러 구현

* refactor: 거리,시간 계산 유틸 common->hub 이동

* feat: jpa auditing 추가

+++ hub/src/main/java/com/klp/hub/hub/presentation/dto/request/hub/UpdateHubRequest.java
@@ -4,8 +4,8 @@ import com.klp.hub.hub.application.command.hub.UpdateHubCommand;
 
 public record UpdateHubRequest(
     String name,
-    Long latitude,
-    Long longitude,
+    Double latitude,
+    Double longitude,
     String address
 ) {
     public UpdateHubCommand toCommand(){
